+++ src/App.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import './App.css';
-import {BrowserRouter as Router, Route} from 'react-router-dom';
-
+import { BrowserRouter as Router, Route } from 'react-router-dom';
 import Home from './containers/Home';
 import Join from './containers/Join';
 import Start from './containers/Start';
